requirements.txt
# Core AI/ML Libraries
openai>=1.35.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# Data & Vector Stores
chromadb>=0.4.0
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.24.0

# Web Framework & APIs
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
aiofiles>=23.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
asyncpg>=0.29.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
pyjwt>=2.8.0

# Async & Queue
asyncio
celery>=5.3.0
redis>=5.0.0

# UI Dependencies
plotly>=5.17.0
streamlit-option-menu>=0.3.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
rich>=13.6.0
typer>=0.9.0
python-multipart>=0.0.6
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.4.0

# Production
gunicorn>=21.2.0
prometheus-client>=0.17.0
